+++ src/main/java/assets/css/custom.css
@@ -1,2 +1,3 @@
 /* Add additional stylesheets below
 -------------------------------------------------- */
+
