+++ src/main/java/models/EnglishGame.java
@@ -1,0 +1,28 @@
+package models;
+
+import java.util.ArrayList;
+
+/**
+ * Created by Christian on 2/16/2016.
+ */
+public class EnglishGame extends Game {
+
+    // Constructor for EnglishGame
+    public EnglishGame() {
+        cols.add(new ArrayList<Card>());
+        cols.add(new ArrayList<Card>());
+        cols.add(new ArrayList<Card>());
+        cols.add(new ArrayList<Card>());
+        error = false;
+    }
+
+    public void buildDeck() {
+        for(int i = 2; i < 15; i++){
+            deck.add(new Card(i,Suit.Clubs));
+            deck.add(new Card(i,Suit.Hearts));
+            deck.add(new Card(i,Suit.Diamonds));
+            deck.add(new Card(i,Suit.Spades));
+        }
+    }
+
+}
