--- src/main/java/views/AcesUp/AcesUpFinal.css
@@ -1,50 +1,0 @@
-body {
-    text-align: justify;
-    background-color: silver;
-}
-
-h1 {
-    text-align: center;
-    font-family: cursive;
-    color: blue;
-}
-
-p {
-    color: white;
-}
-
-text {
-    color: white;
-}
-
-text1 {
-    color: white;
-}
-
-text2 {
-    color: white;
-}
-
-text3 {
-    color: white;
-}
-
-text4 {
-    color: white;
-}
-
-text5 {
-    color: white;
-}
-
-text6 {
-    color: white;
-}
-
-text7 {
-    color: white;
-}
-
-text8 {
-    color: white;
-}
