--- src/main/java/views/AcesUp/AcesUpStartScreen.html
@@ -1,76 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <link href="AcesUpFinal.css" rel="stylesheet">
-    <title>Aces Up!</title>
-</head>
-<body>
-<h1>Aces Up</h1>
-<p>Welcome to Aces Up. To start playing, click on the Start Game button. If you have any questions, try clicking either the Instructions or Rules button.</p>
-<button id="instruct" type="button" onclick="instructionFunction()">Instructions</button>
-<button id="rules" type="button" onclick="rulesFunction()">Rules</button>
-<button id="startEnglishGame" type="button">Start English Game!</button>
-<button id="startSpanishGame" type="button">Start Spanish Game!</button>
-
-<p id="demo"></p>
-<text></text>
-<br>
-<text1></text1>
-<br>
-<text2></text2>
-<br>
-<text3></text3>
-<br>
-<text4></text4>
-<br>
-<text5></text5>
-<br>
-<text6></text6>
-<br>
-<text7></text7>
-<br>
-<text8></text8>
-
-
-<script src="https://ajax.googleapis.com/ajax/libs/jquery/1.12.0/jquery.min.js"></script>
-<script type="text/javascript">
-
-
-        document.getElementById("startEnglishGame").onclick = function () {
-            location.href = "AcesUp.flt.html";
-        }
-        document.getElementById("startSpanishGame").onclick = function () {
-            location.href = "AcesUp.flt.html";
-        }
-        function instructionFunction(){
-            $(document).ready(function(){
-                $("text").text("This is how you play our version of the game:");
-                $("text1").text("Instruction one...");
-                $("text2").text("Instruction two... ");
-                $("text3").text("Instruction three...");
-                $("text4").text("Instruction four...");
-                $("text5").text("Instruction five...");
-                $("text6").text("Instruction six...");
-                $("text7").text("Instruction seven...");
-                $("text8").text("Instruction eight...");
-            });
-        }
-        function rulesFunction(){
-            $(document).ready(function(){
-                $("text").text("These are the rules you should follow:");
-                $("text1").text("1. Deal four cards in a row face up.");
-                $("text2").text("2. If there are two or more cards of the same suit, discard all but the highest-ranked card of that suit.");
-                $("text3").text("3. Repeat step 2 until there are no more pairs of cards with the same suit.");
-                $("text4").text("4. Whenever there are any empty spaces, you may choose the top card of another pile to be put into the empty space. After you do this, go to Step 2.");
-                $("text5").text("5. When there are no more cards to move or remove, deal out the next four cards from the deck face-up onto each pile.");
-                $("text6").text("6. Repeat Step 2, using only the visible, or top, cards on each of the four piles.");
-                $("text7").text("7. When the last four cards have been dealt out and any moves made, the game is over. The fewer cards left in the tableau, the better. To win is to have only the four aces left.");
-                $("text8").text("When the game ends, the number of discarded cards is your score. The maximum score (and thus the score necessary to win) is 48, which means all cards have been discarded except for the four aces, thus the name of the game.");
-            });
-
-        }
-</script>
-
-</body>
-</html>
