+++ src/test/java/models/testGame.java
@@ -17,14 +17,14 @@ public class testGame {
 
     @Test
     public void testGameBuildDeck(){
-        Game g = new Game();
+        EnglishGame g = new EnglishGame();
         g.buildDeck();
         assertEquals(52,g.deck.size());
     }
 
     @Test
     public void testGameInit(){
-        Game g = new Game();
+        EnglishGame g = new EnglishGame();
         g.buildDeck();
         g.shuffle();
         assertNotEquals(2,g.deck.get(0).getValue());
@@ -32,7 +32,7 @@ public class testGame {
 
     @Test
     public void testGameStart(){
-        Game g = new Game();
+        EnglishGame g = new EnglishGame();
         g.buildDeck();
         g.shuffle();
         g.dealFour();
@@ -44,7 +44,7 @@ public class testGame {
 
     @Test
     public void testCustomDeal(){
-        Game g = new Game();
+        EnglishGame g = new EnglishGame();
         g.buildDeck();
         g.customDeal(0,3,6,9);
         assertEquals("2Clubs",g.cols.get(0).get(0).toString());
@@ -55,13 +55,11 @@ public class testGame {
 
     @Test
     public void testRemoveFunction(){
-        Game g = new Game();
+        EnglishGame g = new EnglishGame();
         g.buildDeck();
         g.customDeal(0,3,6,9);
         g.remove(2);
         assertEquals(0,g.cols.get(2).size());
     }
-
-
 
 }
